Add 'Create script' tests for MsSql and Geopackage (#328)
